file reorganized
rename yt-dl folder as resoucre
fixed control packaged name as controller
