+++ Moodify/src/main/java/com/tool/controller/MoodShuffler.java
@@ -1,4 +1,4 @@
-package com.tool.control;
+package com.tool.controller;
 
 import com.tool.model.Node;
 import com.tool.model.DoublyLinkedList;
