+++ Moodify/src/main/java/com/tool/controller/TestMoodShuffle.java
@@ -1,4 +1,4 @@
-package com.tool.control;
+package com.tool.controller;
 
 import com.tool.model.DoublyLinkedList;
